feat(frigo): On frigo orders, publish some info on kelder mqtt.
The messages are sent via websocket over the LAN. This would break the
secure context of the frigo page, but we can send them to localhost
instead where a `socat` daemon proxies us to the broker.

The actual content of the messages might be changed, the products are
now just database dumps as json.